Project/djangoProject/requirements.txt
django==3.2.14
djangorestframework==3.13.1
